+++ Backend/bank-test/src/main/java/com/consultec/org/banktest/repository/entity/TransferenciaEntity.java
@@ -3,6 +3,7 @@ package com.consultec.org.banktest.repository.entity;
 import jakarta.persistence.*;
 
 import java.util.Date;
+import java.util.List;
 import java.util.UUID;
 
 @Entity
@@ -11,19 +12,13 @@ public class TransferenciaEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
-    private double monto;
     private Date fechaTransferencia;
-    @OneToOne(mappedBy = "transferencia")
-    private MovimientoEntity movimientos;
-    @ManyToOne
-    @JoinColumn(name = "numCuentaDest")
-    private CuentaEntity numeroCuenta;
+    @OneToMany(mappedBy = "transferencia")
+    private List<MovimientoEntity> movimientos;
 
-    public TransferenciaEntity(UUID id, double monto, Date fechaTransferencia, CuentaEntity numeroCuenta) {
+    public TransferenciaEntity(UUID id, Date fechaTransferencia) {
         this.id = id;
-        this.monto = monto;
         this.fechaTransferencia = fechaTransferencia;
-        this.numeroCuenta = numeroCuenta;
     }
 
     public UUID getId() {
@@ -34,14 +29,6 @@ public class TransferenciaEntity {
         this.id = id;
     }
 
-    public double getMonto() {
-		return monto;
-	}
-
-	public void setMonto(double monto) {
-		this.monto = monto;
-	}
-
 	public Date getFechaTransferencia() {
         return fechaTransferencia;
     }
@@ -50,11 +37,4 @@ public class TransferenciaEntity {
         this.fechaTransferencia = fechaTransferencia;
     }
 
-	public CuentaEntity getNumeroCuenta() {
-		return numeroCuenta;
-	}
-
-	public void setNumeroCuenta(CuentaEntity numeroCuenta) {
-		this.numeroCuenta = numeroCuenta;
-	}
 }
